>= 0.23.4 use less doubles in navigation interpolator

+++ src/pymodule/c_navigationinterpolatorlatlon.cpp
@@ -77,7 +77,7 @@ void init_c_NavigationInterpolatorLatLon(py::module& m)
                  set_sensor_configuration),
              py::arg("sensor_configuration"))
         .def("add_target",
-             py::overload_cast<const std::string&, double, double, double, double, double, double>(
+             py::overload_cast<const std::string&, float, float, float, float, float, float>(
                  &NavigationInterpolatorLatLon::add_target),
              DOC(themachinethatgoesping, navigation, I_NavigationInterpolator, add_target),
              py::arg("target_id"),
@@ -116,7 +116,7 @@ void init_c_NavigationInterpolatorLatLon(py::module& m)
 
         // set depth data
         .def("set_data_depth",
-             py::overload_cast<const std::vector<double>&, const std::vector<double>&>(
+             py::overload_cast<const std::vector<double>&, const std::vector<float>&>(
                  &NavigationInterpolatorLatLon::set_data_depth),
              DOC(themachinethatgoesping, navigation, I_NavigationInterpolator, set_data_depth),
              py::arg("timestamp"),
@@ -125,8 +125,8 @@ void init_c_NavigationInterpolatorLatLon(py::module& m)
         // set attitude data (no yaw)
         .def("set_data_attitude",
              py::overload_cast<const std::vector<double>&,
-                               const std::vector<double>&,
-                               const std::vector<double>&>(
+                               const std::vector<float>&,
+                               const std::vector<float>&>(
                  &NavigationInterpolatorLatLon::set_data_attitude),
              DOC(themachinethatgoesping, navigation, I_NavigationInterpolator, set_data_attitude),
              py::arg("timestamp"),
@@ -135,7 +135,7 @@ void init_c_NavigationInterpolatorLatLon(py::module& m)
 
         // set data compass
         .def("set_data_heading",
-             py::overload_cast<const std::vector<double>&, const std::vector<double>&>(
+             py::overload_cast<const std::vector<double>&, const std::vector<float>&>(
                  &NavigationInterpolatorLatLon::set_data_heading),
              DOC(themachinethatgoesping, navigation, I_NavigationInterpolator, set_data_heading),
              py::arg("timestamp"),
@@ -176,14 +176,14 @@ void init_c_NavigationInterpolatorLatLon(py::module& m)
             "interpolator_depth",
             &NavigationInterpolatorLatLon::interpolator_depth,
             [](NavigationInterpolatorLatLon& self,
-               const themachinethatgoesping::tools::vectorinterpolators::LinearInterpolator<double, double>&
+               const themachinethatgoesping::tools::vectorinterpolators::LinearInterpolator<double, float>&
                    interpolator) { self.interpolator_depth() = interpolator; },
             DOC(themachinethatgoesping, navigation, I_NavigationInterpolator, interpolator_depth))
         .def_property(
             "interpolator_attitude",
             &NavigationInterpolatorLatLon::interpolator_attitude,
             [](NavigationInterpolatorLatLon& self,
-               const themachinethatgoesping::tools::vectorinterpolators::SlerpInterpolator<double, double>&
+               const themachinethatgoesping::tools::vectorinterpolators::SlerpInterpolator<double, float>&
                    interpolator) { self.interpolator_attitude() = interpolator; },
             DOC(themachinethatgoesping,
                 navigation,
@@ -193,7 +193,7 @@ void init_c_NavigationInterpolatorLatLon(py::module& m)
             "interpolator_heading",
             &NavigationInterpolatorLatLon::interpolator_heading,
             [](NavigationInterpolatorLatLon& self,
-               const themachinethatgoesping::tools::vectorinterpolators::SlerpInterpolator<double, double>&
+               const themachinethatgoesping::tools::vectorinterpolators::SlerpInterpolator<double, float>&
                    interpolator) { self.interpolator_heading() = interpolator; },
             DOC(themachinethatgoesping, navigation, I_NavigationInterpolator, interpolator_heading))
         .def("merge",
